+++ packages/xrpc/test/src/const_test.dart
@@ -11,5 +11,6 @@ import 'package:xrpc/src/const.dart';
 void main() {
   test('default service', () {
     expect(defaultService, 'bsky.social');
+    expect(defaultStreamService, 'bsky.network');
   });
 }