+++ packages/atproto/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.8.1
 
 - Upgraded `atproto_core`. ([#999](https://github.com/myConsciousness/atproto.dart/issues/999))
+- Fixed a bug that service names were not specified correctly when using stream endpoints.
+- Added `.streamService` property on `ATProto`. Defaults to `bsky.network`.
 
 ## v0.8.0