+++ packages/xrpc/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.4.1
 
 - Added `.post` function. ([#999](https://github.com/myConsciousness/atproto.dart/issues/999))
+- Use `bsky.network` as default when service is missing for `.subscribe` function.
 
 ## v0.4.0